requirements.txt
pytest==9.0.2
requests==2.32.5
selenium==4.41.0
allure-pytest==2.15.3
python-dotenv==1.2.1